+++ lib/screens/home/home_logging_cards.dart
@@ -512,6 +512,7 @@ class HomeLoggingCards extends StatelessWidget {
                     }(),
                     showButtons: !isGuest,
                     onAdd: onShowWeightSheet,
+                    onChart: () => context.push('/weight/analytics'),
                   ),
                 ),
               ],